moved palettes to their own folder
