+++ src/components/palette/Palette.js
@@ -1,7 +1,6 @@
 import React, { useState, useContext } from "react";
 import { useQuery } from "react-query";
 import styled from "styled-components";
-import { Swatch } from "./reusable/Swatch.js";
 import {
   H4,
   Col,
@@ -12,14 +11,16 @@ import {
   CardFooter,
   Tooltip,
 } from "@bootstrap-styled/v4";
-import { ColorContext } from "./ColorProvider.js";
+import { Swatch } from "../reusable/Swatch.js";
+import { CopyButton } from "../reusable/CopyButton.js";
+import { ColorContext } from "../ColorProvider.js";
 import { PaletteContext } from "./PaletteProvider.js";
-import { DEFAULT_PALETTE_MINIMAL } from "../utils/color.js";
-import { isNobody } from "../utils/auth.js";
-import { CopyButton } from "./reusable/CopyButton.js";
+import { DEFAULT_PALETTE_MINIMAL } from "../../utils/color.js";
+import { isNobody } from "../../utils/auth.js";
+import { KEYS } from "../../utils/query.js";
 
 export const Palette = ({ ...props }) => {
-  const { color, setColor, KEYS } = useContext(ColorContext);
+  const { color, setColor } = useContext(ColorContext);
   const { getPalette } = useContext(PaletteContext);
   const [name, setName] = useState("default");
   // if negative, all clean
@@ -32,11 +33,9 @@ export const Palette = ({ ...props }) => {
     [KEYS.CURRENT_PALETTE, name],
     () => getPalette(name),
     {
-      onSuccess: () => {
+      onSuccess: (data) => {
         setDirtyColor(-1);
-        setColors(
-          palette.data.colors.map((colorObj) => colorObj.color.rgb_hex)
-        );
+        setColors(data.colors.map((colorObj) => colorObj.color.rgb_hex));
       },
       initialData: DEFAULT_PALETTE_MINIMAL,
       keepPreviousData: true,
